refactor: lot 8b — découper placementDeckBuilder.ts (523 l.) en 3 fichiers
- placementDeckBuilder.types.ts : 6 interfaces data + LEGAL_TEXT (exportés)
- placementDeckBuilder.helpers.ts : formatters, params builders, slide spec builders, projection builders
- placementDeckBuilder.ts : orchestrateur fin (~115 l.) — importe et ré-exporte les types, garde uniquement buildPlacementStudyDeck
